[General] Add clang-format configuration
I am adding a .clang-format file to automate keeping the coding style
consistent.

This should help make pull requests from new contributors (like myself)
fit in with the existing codebase.

My intention with this clang-format file is to be close to the style
used by visual studio. Any recommendations on how to get closer to that
(assuming that is a good goal) would be much appreciated.