Draft for compiling NovusCore on macos
This is a draft commit for building NovusCore on macos, working towards
building on linux. To keep things simple I am only building the
authserver and common.

So far the build is a bit janky, but it works and Nix has been able to
connect from his client to the authserver.