[General] Fix worldnode and tools build for macos
The WorldNode and data extrator in tools use std::filesystem which is
not available by default in llvm 8, which is the latest version of llvm
that you can get via homebrew.

There was a fix in place but that compiler flag only seems to work in linux.

This patch allows us to compile projects that use std::filesystem under
linux and macos.