[Common] Use DebugHandler in DatabaseHandler
It looks like the DatabaseHandler was created before the DebugHandler.
This commit swaps out the standard logs in the db handler out for the
DebugHandler.

Previously the applications would continue to run if they could not
connect to the database. This has been changed to a fatal error as we
cannot continue past this point.