[WorldNode] Surpress clang compiler warnings
ObjectGuid: It seems that clang raises a warning when you don't include
an enum value in a switch statement. Adding a default seems to fix this.

UpdateData: Clang found the usage of ternary and arithmatic without
explicit parentheses. The default buffer size is fine for now so
removing the code is the neatest course of action.