Fix clang compiler warning on macos (#17)
* [Common] Surpress clang compiler warnings in DebugHandler

Unfortunately we cannot currenlty figure out a way to sort out this
compiler warning without losing the simplicity of the DebugHandler.

* [WorldNode] Surpress clang compiler warnings

ObjectGuid: It seems that clang raises a warning when you don't include
an enum value in a switch statement. Adding a default seems to fix this.

UpdateData: Clang found the usage of ternary and arithmatic without
explicit parentheses. The default buffer size is fine for now so
removing the code is the neatest course of action.

* [RealmServer] Surpress clang compiler warnings

Previously clang would raise the following warning for the changed line
of code:

/NovusCore/projects/realmserver/Connections/RealmConnection.cpp:221:26: warning: '&&' within '||' [-Wlogical-op-parentheses]

* [Deps] Suppress clang warnings from libmpq